FileクラスをFiledetailクラスに命名変更・detailメソッドをto_sメソッドのオーバーライドに変更
